05-data-platforms/bruin/my-taxi-pipeline/pipeline/assets/ingestion/requirements.txt
# Python dependencies for NYC taxi ingestion asset.
# Pin versions for reproducibility.
# Docs: https://getbruin.com/docs/bruin/assets/python

pandas>=2.0.0
pyarrow>=14.0.0
python-dateutil>=2.8.0
